+++ 2026/04_Webアプリケーション開発基礎/assignment/backend/README.md
@@ -1,0 +1,15 @@
+# (BE) アーキテクチャの選定について
+
+アーキテクチャはレイヤード・アーキテクチャを採用。
+プレゼンテーション、ビジネスロジック、データの三つに分類して、依存関係と役割を明確化した。
+
+またエントリーポイントと内部の処理はディレクトリで分離した。（自分は余り Go を触ったことがないのですが、このディレクトリ構造が多いようなので）
+
+```
+backend/
+├── cmd/server/ # アプリケーションのエントリーポイント
+├── internal/ # プライベート API（外部からインポート不可）
+│ ├── handlers/ # HTTP 層 - プレゼンテーション
+│ ├── models/ # ドメインモデル - ビジネス
+│ └── store/ # データアクセス層 - インフラ
+```
